added userinfo icon and created a component for it
