--- nba-front-end/src/components/Menu/UserInformation/UserInformation.css
@@ -1,20 +1,0 @@
-.blue-circle{
-    margin-right: 18px;
-    justify-content: center;
-    display: flex;
-    padding-top:10px;    
-    width: 45px;
-    height: 45px;
-    background-color: #1976d2;
-    color: white;
-    font-weight: bold;
-    border-radius: 50px;
-}
-
-.full-email{
-    font-size: smaller;
-    overflow: hidden;
-    text-overflow: ellipsis; 
-    width:min-content; 
-    width:150px;
-}
