+++ nba-front-end/src/components/Menu/UserInformationMenu/UserInformationMenu.css
@@ -1,0 +1,25 @@
+.user-logo{
+    justify-content: center;
+    display: flex;
+    width: 50px;
+    height: 50px;
+    align-items: center;
+    background-color: white;
+    color: #1976d2 !important;
+    border-color: transparent;
+    font-weight: bold;
+    border-radius: 50%;
+    font-size: 25px;   
+}
+
+.user-logo:hover {
+    background-color: #336ca5ba;
+    color: white !important;
+}
+
+@media only screen and (max-width: 599px) {
+    .user-logo{
+        width: 40px;
+        height: 40px; 
+    }
+  }
